[Fix] 레시피 재료 생성자 타입 수정

+++ src/main/java/com/inmyhand/refrigerator/recipe/domain/entity/RecipeIngredientEntity.java
@@ -35,7 +35,7 @@ public class RecipeIngredientEntity {
     @JsonIgnore
     private RecipeInfoEntity recipeInfoEntity;
 
-    public RecipeIngredientEntity(String ingredientName, Integer ingredientQuantity,  String ingredientUnit, RecipeInfoEntity recipeInfoEntity) {
+    public RecipeIngredientEntity(String ingredientName, double ingredientQuantity,  String ingredientUnit, RecipeInfoEntity recipeInfoEntity) {
         this.ingredientName = ingredientName;
         this.ingredientQuantity = ingredientQuantity;
         this.ingredientUnit = ingredientUnit;
